pyproject.toml
[project]
name = "blob-store"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Hariom Narang", email = "hariom.2711@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.19.0",
    "azure-storage-blob>=12.24.0",
    "boto3>=1.35.68",
    "google-cloud-storage>=2.18.2",
    "pydantic>=2.10.1",
    "typing-extensions>=4.12.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "boto3-stubs>=1.35.68",
    "ipykernel>=6.29.5",
    "ipython>=8.29.0",
    "pytest>=8.3.3",
]
